+++ circuits/types/field/src/double.rs
@@ -57,7 +57,7 @@ mod tests {
     use snarkvm_circuits_environment::Circuit;
     use snarkvm_utilities::{test_rng, UniformRand};
 
-    const ITERATIONS: usize = 10_000;
+    const ITERATIONS: u64 = 10_000;
 
     fn check_double(name: &str, mode: Mode) {
         for _ in 0..ITERATIONS {
