+++ circuits/types/field/src/from_bits.rs
@@ -132,15 +132,9 @@ mod tests {
     use snarkvm_circuits_environment::Circuit;
     use snarkvm_utilities::{test_rng, UniformRand};
 
-    const ITERATIONS: usize = 100;
+    const ITERATIONS: u64 = 100;
 
-    fn check_from_bits_le(
-        mode: Mode,
-        num_constants: usize,
-        num_public: usize,
-        num_private: usize,
-        num_constraints: usize,
-    ) {
+    fn check_from_bits_le(mode: Mode, num_constants: u64, num_public: u64, num_private: u64, num_constraints: u64) {
         for i in 0..ITERATIONS {
             // Sample a random element.
             let expected: <Circuit as Environment>::BaseField = UniformRand::rand(&mut test_rng());
@@ -160,7 +154,7 @@ mod tests {
             });
 
             // Add excess zero bits.
-            let candidate = vec![given_bits, vec![Boolean::new(mode, false); i]].concat();
+            let candidate = vec![given_bits, vec![Boolean::new(mode, false); i as usize]].concat();
 
             Circuit::scope(&format!("Excess {} {}", mode, i), || {
                 let candidate = Field::<Circuit>::from_bits_le(&candidate);
@@ -178,13 +172,7 @@ mod tests {
         }
     }
 
-    fn check_from_bits_be(
-        mode: Mode,
-        num_constants: usize,
-        num_public: usize,
-        num_private: usize,
-        num_constraints: usize,
-    ) {
+    fn check_from_bits_be(mode: Mode, num_constants: u64, num_public: u64, num_private: u64, num_constraints: u64) {
         for i in 0..ITERATIONS {
             // Sample a random element.
             let expected: <Circuit as Environment>::BaseField = UniformRand::rand(&mut test_rng());
@@ -204,7 +192,7 @@ mod tests {
             });
 
             // Add excess zero bits.
-            let candidate = vec![vec![Boolean::new(mode, false); i], given_bits].concat();
+            let candidate = vec![vec![Boolean::new(mode, false); i as usize], given_bits].concat();
 
             Circuit::scope(&format!("Excess {} {}", mode, i), || {
                 let candidate = Field::<Circuit>::from_bits_be(&candidate);
