+++ circuits/types/group/src/equal.rs
@@ -49,7 +49,7 @@ mod tests {
     use snarkvm_circuits_environment::Circuit;
     use snarkvm_utilities::{test_rng, UniformRand};
 
-    const ITERATIONS: usize = 100;
+    const ITERATIONS: u64 = 100;
 
     #[test]
     fn test_is_equal() {
