+++ circuits/types/integers/src/compare.rs
@@ -102,7 +102,7 @@ mod tests {
 
     use std::ops::RangeInclusive;
 
-    const ITERATIONS: usize = 100;
+    const ITERATIONS: u64 = 100;
 
     fn check_compare<I: IntegerType>(name: &str, first: I, second: I, mode_a: Mode, mode_b: Mode) {
         let a = Integer::<Circuit, I>::new(mode_a, first);
