+++ circuits/types/integers/src/mul_checked.rs
@@ -187,7 +187,7 @@ mod tests {
 
     use core::{ops::RangeInclusive, panic::RefUnwindSafe};
 
-    const ITERATIONS: usize = 32;
+    const ITERATIONS: u64 = 32;
 
     fn check_mul<I: IntegerType + RefUnwindSafe>(name: &str, first: I, second: I, mode_a: Mode, mode_b: Mode) {
         let a = Integer::<Circuit, I>::new(mode_a, first);
